requirements.txt
amqp==5.0.9
asgiref==3.4.1
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.7
django-cors-headers==3.10.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-filter==21.1
django-js-asset==1.2.2
django-models-logging==1.0.8
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
ipython==7.30.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.2
kombu==5.2.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.3
parso==0.8.3
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
pkgconfig==1.5.5
prompt-toolkit==3.0.24
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
PyNaCl==1.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pyvips==2.1.16
redis==4.1.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.0
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
